liquidctl: Reduce noise overall
This system is currently running moderate to heavy jobs most of the
time, and I'm near it even more than usual.

So sacrifice a tiny bit of cooling (~2°C CPU/GPU for mixed loads;
possibly a little more on the GPU during GPU-only loads) in order to
bring the noise level down, especially in the higher pitched range.