nvim: start recording lazy-lock.json again
Even if it doesn't show the full picture (e.g. the Mason packages), and
that I usually don't set specific versions because Lazy isn't
particularly nice to use with anything other than U:update (all),
this can be helpful to track down bugs and make sanity checks in the
process.